fix: move types 🔨
